Split data stored in kvstore into chunks

+++ src/storage/filesystem.ts
@@ -3,10 +3,10 @@ import PATH from "path";
 
 class FS {
     async set(key: string, value: string) {
-        if (!fs.existsSync(PATH.dirname(key))) {
+        if (value && value !== '""' && !fs.existsSync(PATH.dirname(key))) {
             fs.mkdirSync(PATH.dirname(key), { recursive: true });
         }
-        if (!value || value === "{}") {
+        if (!value || value === '""') {
             return await this.remove(key);
         }
         return fs.writeFileSync(key, value);
@@ -14,13 +14,27 @@ class FS {
 
     async get(key: string) {
         if (!fs.existsSync(key)) {
-            return "{}";
+            return '""';
         }
         return fs.readFileSync(key);
     }
 
     async remove(key: string) {
-        return fs.unlinkSync(key);
+        if (!fs.existsSync(key)) {
+            return;
+        }
+        fs.unlinkSync(key);
+        while (fs.readdirSync(PATH.dirname(key)).length === 0) {
+            key = PATH.dirname(key);
+            fs.rmdirSync(key);
+        }
+    }
+
+    async list(key: string) {
+        if (!fs.existsSync(key)) {
+            return [];
+        }
+        return fs.readdirSync(key);
     }
 }
